generated_code/fixtures/tests/test_parseunittestcoverage_fixture.py
from code_autoeval.llm_model.utils.logging_statements.logging_statements import LoggingStatements
from code_autoeval.llm_model.utils.logging_statements.logging_statements import LoggingStatements
import pytest
from generated_code.fixtures.fixtures.parseunittestcoverage_fixture import fixture_mock_parseunittestcoverage
from code_autoeval.llm_model.utils.extraction.parse_unit_test_coverage import ParseUnitTestCoverage
from pathlib import Path
from typing import Any, Dict
def test_mock_parseunittestcoverage(fixture_mock_parseunittestcoverage):
    assert isinstance(fixture_mock_parseunittestcoverage, ParseUnitTestCoverage)
    assert hasattr(fixture_mock_parseunittestcoverage, 'coverage_result')
    assert hasattr(fixture_mock_parseunittestcoverage, 'file_path')
    assert isinstance(fixture_mock_parseunittestcoverage.file_path, Path)
    assert hasattr(fixture_mock_parseunittestcoverage, 'test_file_path')
    assert isinstance(fixture_mock_parseunittestcoverage.test_file_path, Path)
    assert hasattr(fixture_mock_parseunittestcoverage, 'absolute_path_from_root')
    assert isinstance(fixture_mock_parseunittestcoverage.absolute_path_from_root, Path)
    assert hasattr(fixture_mock_parseunittestcoverage, 'unique_imports_dict')
    assert isinstance(fixture_mock_parseunittestcoverage.unique_imports_dict, Dict)
    assert hasattr(fixture_mock_parseunittestcoverage, 'run_parse_unit_test_cov')
    assert isinstance(fixture_mock_parseunittestcoverage.run_parse_unit_test_cov, classmethod)
    assert isinstance(fixture_mock_parseunittestcoverage, LoggingStatements)